usersession: move userd package to usersession/userd
