+++ usersession/userd/ui/zenity_test.go
@@ -27,7 +27,7 @@ import (
 	. "gopkg.in/check.v1"
 
 	"github.com/snapcore/snapd/testutil"
-	"github.com/snapcore/snapd/userd/ui"
+	"github.com/snapcore/snapd/usersession/userd/ui"
 )
 
 func Test(t *testing.T) { TestingT(t) }
